# openapi.json
{"openapi":"3.1.0","info":{"title":"Fonio FDE Lab","description":"Mock integration API for HausPilot Immobilienverwaltung GmbH.","version":"0.1.0"},"paths":{"/customers/by-phone/{phone}":{"get":{"tags":["customers"],"summary":"Find Customer By Phone","operationId":"find_customer_by_phone_customers_by_phone__phone__get","parameters":[{"name":"phone","in":"path","required":true,"schema":{"type":"string","title":"Phone"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerLookupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/open-tickets":{"get":{"tags":["customers"],"summary":"Get Customer Open Tickets","operationId":"get_customer_open_tickets_customers__customer_id__open_tickets_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"},"title":"Response Get Customer Open Tickets Customers  Customer Id  Open Tickets Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/properties/{property_id}":{"get":{"tags":["properties"],"summary":"Get Property","operationId":"get_property_properties__property_id__get","parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","title":"Property Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Property"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tickets":{"post":{"tags":["tickets"],"summary":"Create Ticket","operationId":"create_ticket_tickets_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ticket"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"Customer":{"properties":{"id":{"type":"string","title":"Id"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"phone":{"type":"string","title":"Phone"},"property_id":{"type":"string","title":"Property Id"},"unit":{"type":"string","title":"Unit"},"property_manager_id":{"type":"string","title":"Property Manager Id"}},"type":"object","required":["id","first_name","last_name","phone","property_id","unit","property_manager_id"],"title":"Customer"},"CustomerLookupResponse":{"properties":{"status":{"$ref":"#/components/schemas/MatchStatus"},"count":{"type":"integer","title":"Count"},"customers":{"items":{"$ref":"#/components/schemas/Customer"},"type":"array","title":"Customers"}},"type":"object","required":["status","count","customers"],"title":"CustomerLookupResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MatchStatus":{"type":"string","enum":["not_found","unique","ambiguous"],"title":"MatchStatus"},"Property":{"properties":{"id":{"type":"string","title":"Id"},"address":{"type":"string","title":"Address"},"property_manager_id":{"type":"string","title":"Property Manager Id"},"emergency_contact_id":{"type":"string","title":"Emergency Contact Id"}},"type":"object","required":["id","address","property_manager_id","emergency_contact_id"],"title":"Property"},"Ticket":{"properties":{"customer_id":{"type":"string","title":"Customer Id"},"property_id":{"type":"string","title":"Property Id"},"category":{"type":"string","maxLength":100,"minLength":1,"title":"Category"},"description":{"type":"string","maxLength":2000,"minLength":1,"title":"Description"},"priority":{"$ref":"#/components/schemas/TicketPriority"},"id":{"type":"string","title":"Id"},"status":{"$ref":"#/components/schemas/TicketStatus"}},"type":"object","required":["customer_id","property_id","category","description","priority","id","status"],"title":"Ticket"},"TicketCreate":{"properties":{"customer_id":{"type":"string","title":"Customer Id"},"property_id":{"type":"string","title":"Property Id"},"category":{"type":"string","maxLength":100,"minLength":1,"title":"Category"},"description":{"type":"string","maxLength":2000,"minLength":1,"title":"Description"},"priority":{"$ref":"#/components/schemas/TicketPriority"}},"type":"object","required":["customer_id","property_id","category","description","priority"],"title":"TicketCreate"},"TicketPriority":{"type":"string","enum":["low","medium","high"],"title":"TicketPriority"},"TicketStatus":{"type":"string","enum":["open","in_progress","closed"],"title":"TicketStatus"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}